delete column rs4917623
Removed intronic variant in *18 and *19, rs4917623, from spreadsheet.
This variant is in about 40% of the population and pubs show it can be
found on the *1 and *2 background.  Having it only in the definition
for *18 and *19 forces many St. Jude wgs calls into no calls because
the intronic snp is present and the exonic defining snps are not.